fix(plugin-search): Render error on custom UI component
